memory optimisation for Z,U and Sectors variables.

+++ StRoot/StHbtMaker/Infrastructure/TpcLocalTransform.cc
@@ -9,7 +9,7 @@
 
 
 int TpcLocalTransform(StThreeVectorD& aPoint, int& aSector, int& aRow, 
-		      double& aU, double& aPhi){
+		      float& aU, double& aPhi){
   static int tNPadAtRow[45]={
   88,96,104,112,118,126,134,142,150,158,166,174,182,
   98,100,102,104,106,106,108,110,112,112,114,116,118,120,122,122,
